backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv

# Scraping
beautifulsoup4
requests

# Summarization (if using Hugging Face Transformers)
transformers
torch

# Optional utilities
lxml
newspaper3k
